ui: update dashboard page
